fix(engine): read hash_bytes' tail byte as unsigned
The tail loop read through a SIGNED char, so a byte >= 0x80 became
0xFFFFFFFFFFFFFF00 | byte and the `or` set every bit above bit 7 -- erasing
everything the higher tail indices had already contributed. Over all 65536
two-byte inputs that leaves 32896 distinct digests instead of 65536, and
{0x80,0x02,0x03} and {0x80,0x02,0x04} collide outright. Reference
MurmurHash64A reads through `const unsigned char*`; upstream's `misc.cpp:481`
does not, and inherits the defect.

THIS TREE REPRODUCED IT ON PURPOSE, and the comment saying so was wrong about
why. It claimed zero-extending "would silently produce a different digest",
which is true and irrelevant: the only readers are the three `*_content_hash`
functions, which upstream uses to name a shared-memory segment and which
NOTHING here calls. The net reader and writer compare the u32 ARCHITECTURE
hashes -- `nnue_network_hash_value` and friends, built by `combine_hash` over
dimensions -- which never reach this function. Verified by tracing every
reference: `nnue_hash_bytes`, `nnue_hash_combine` and all three content hashes
have zero callers in src/, tests/ and tools/. So the comment was instructing
future readers to preserve a defect for a consumer that does not exist, which
is the more expensive half of leaving it alone.

Gated by census rather than by spot value, because the symptom is a collapsed
RANGE and no existing gate can see a digest nothing reads:

  over all 65536 two-byte inputs   32896 distinct -> 65536 distinct
  {80,02,03} vs {80,02,04}         equal -> different

The third assertion is the one that says this is a fix to the sign extension and
not to the algorithm: a tail with no high bit must hash to reference
MurmurHash64A's value for the same bytes. That constant (0xf187641bcbc5b011 for
{01,02,03}) was computed from an independent Python implementation of the
reference, not read off this one -- pinning what mcfish prints would only
photograph current behaviour. It matches, which also confirms the rest of the
algorithm was already right.

  ./build.sh negative-control test
    test -- hash_bytes sign-extends its tail    ok  went red (exit 1)
  ./build.sh test      1620913 checks, 0 failures            PASSED
  ./build.sh parity    === parity: all gates passed ===      exit 0
    signature 2884956 unmoved -- the bench never reaches this function

Closes the last of the twenty upstream defects that was still open here and
fixed on refish (`bf567c41`).

+++ src/engine/eval/nnue/nnue_hash.c
@@ -39,12 +39,26 @@ uint64_t nnue_hash_bytes(const uint8_t *data, size_t len) {
         size_t i = len & 7;
         while (i > 0) {
             i -= 1;
-            // SIGN-extend. Upstream's hash_bytes takes `const char*` and does
-            // `u64(end[i])` (misc.cpp:481); plain char is SIGNED on x86-64, so a
-            // tail byte >= 0x80 fills the high bits with ones. Zero-extending
-            // here would silently produce a different digest for any input whose
-            // length is not a multiple of 8 and whose tail has the high bit set.
-            k = (k << 8) | (uint64_t) (int64_t) (int8_t) data[tail + i];
+            // ZERO-extend, as reference MurmurHash64A does by reading through a
+            // `const unsigned char*`.
+            //
+            // Upstream sign-extends by accident: `hash_bytes` takes a `const char*`
+            // and does `u64(end[i])` (misc.cpp:481), and plain char is SIGNED here,
+            // so a tail byte >= 0x80 becomes 0xFFFFFFFFFFFFFF00 | byte -- and the
+            // `or` then sets every bit above bit 7, ERASING everything already
+            // accumulated from the higher tail indices. Over all 65536 two-byte
+            // inputs that leaves 32896 distinct digests instead of 65536, and
+            // {0x80,0x02,0x03} and {0x80,0x02,0x04} collide outright.
+            //
+            // Reproducing it was deliberate while it looked like a compatibility
+            // requirement. It is not one: the only readers of this digest are the
+            // three `*_content_hash` functions, which upstream uses to name a
+            // shared-memory segment and which NOTHING in this tree calls -- the net
+            // reader and writer compare the u32 ARCHITECTURE hashes, built by
+            // `combine_hash` over dimensions, which never reach this function. So
+            // the old comment was instructing future readers to preserve a defect
+            // for a consumer that does not exist.
+            k = (k << 8) | (uint64_t) data[tail + i];
         }
         h ^= k;
         h *= m;
